[IMP] web, account: remove serialized context from report URL
Previously while printing few reports (mostly the ones that are printed
using wizard), some data required for printig the report was aslo being
passed by serializing context and adding it to report url from action manager,
which should not be the case.

This means that all the reports should utilize only provided key-value
data and should NOT use values from context.

This commit improves the behavior by passing required values into data
only and removing context from the report url.

Note: Context was also used to decide the orientation of a report, so from
this commit, we will pass key-value pair in the data dictionary for doing
the same. If 'orientation' key is not specified in data, by default report
will have portrait orientation. For printing a landscpae report, simply pass
'orientation' key with value as 'landscape'.

Related to task 21893