[FIX] website_sale_coupon: make test determinist
avoid race condition when refresh of /shop finish after the search click.

Now we start on the search page directly, that make tour faster and that
don't remove any usefull test, since the purpose of this test is to test
coupon and not the search view already tested by others tests.