[IMP] mrp: Added an option in the settings to show the timer on the work order's tablet screen
Purpose
=======

When a user opens a WO, the time starts to be recorded automatically. However, users don't always notice it and so they forget to click on pause when they stop working.
Thus, a timer (optional) should be available on the WO tablet view so that the user understands what happens.

Spec
======
Removed field 'duration' from the form view of WO.
Added an option in the settings to visible/disable the timer on the work order's tablet screen.

Related to Task: 1896784