[REF] mail, various : make message_partner_info_from_emails from mail thread private
Purpose: make some methods from mail.thread mixin private and update their
naming if necessary. Less publicly available mixin methods from mail.thread
means more control on information and access checks.

In this commit we put message_partner_info_from_emails as private as it is
not intended to be used externally. Instead we define a controller used
in chatter that grants access to this method.

Some addons are updated accordingly to the method change.

Related to task ID 1911679
Linked to PR #29483