[IMP] sale_stock, sale_mrp: Be able to cancel an order if the products have been returned.
Purpose of this task :
Cancel the sale order if all the delivered products have been returned.

After this commit:
If all the products which we delivered to the customer, and customer returned all the products, then user is able to cancel the sale order.
For example:
User create a sale order of quantity 3 and deliverd 3 quantity to customer, but for some reason customer returned all the 3 quantity,
but in our system sale order state is 'Sale Order', and user is not able to cancle it,
though it shows that our stock is having the same status before this sales order place.
So the scenario is, when sale order's delivered quantity and returned quantity are same then user is able to cancel the sales order.

This commit is related to task id : 1919761.