[ADD] web_editor, website: add a video as a background
This commit adds facility to add a video as a background
on the website for any snippet which can already have a
background image/color.

"Background image" option is replaced by "Background" and
now it will open both image/video tabs for background.

"Muted" option is added for video options and all other
options are checked by default.
"Opacity" with scroll is added.

task- 1920630