[IMP] account,point_of_sale: robustness
prevent editing in pos_config with active session:
    active
    module_pos_restaurant
    journal_ids
    sequence_ids
    pickint_type_id
    journal_id

prevent deletion and deactivation of models used
in pos session:
    account_account
    account_journal
    ir_sequence
    product_template
    res_partner
    res_users
    stock_picking_type

If locally stored session information does not match with the available
database information, show a popup windows explaining there is a
conflict in the data at hand. The popup gives the user the oportunity to
save the local storage data and clear the local storage so they can
continue selling.

Task id: 1879971