[REF] stock_account: valuation layers
Standard cost method:
**************************************************
The update of the standard price logic was slightly changed: before, the
code looped over all quants and created an accounting entry for every
location where some quants were available. It generated some noise if
quants were available at a lot of locations and it was also wrong by not
considering that a quant could be valued or not. We now create a single
new valuation layer and we consider the valued quantity instead of
blindly the quantity in stock.

TODO: move the journal to the category and remove the update std price
wizard

AVCO cost method:
**************************************************
- Make `product_price_update_before_done` private
- Use the quantity valued and not quantity available when computing the
  new AVCO
- use the `_is_in` helper instead of hardcoding a check on the move's
  location source

TODO: AVCO vacuum

FIFO cost method:
**************************************************
implemente me

Change of cost Method
**************************************************
Before, we always computed the fifo stack even in standard and AVCO.
This was used in order to switch to the fifo cost method anytime. It was
wrong because no fifo vaucuum ran, thus the stack wasn't the same s if
the fifo cost method was used normally. When switch to avco, we
recomputed an AVCO price on the fly. It was problematic because we
didn't know if the user changed its cost method and it made really
difficul to understand the value of the standard price.

We now implemented this logic: we don't compute the fifo stack in fifo
and avco. We only compute the avco when in AVCO, and so on. If the user
changes its cost method, we create a out valuation layer to empty the
stock of the valued quantity, then we create an in valuation layer with
the valued quantity, filling remaining_qty if we change to fifo and using
the current standard price when we change to avco.

The accounting impact of the cost method changes are neutrol except fifo
to standard and fifo to avco

TODO:
    * create an empty valuation layer to track the changes
    * description on valuaiton layer change

edit done move line: dropshipped is now supported (but not tested)

vacuum is ran at each receipt and no more in the scheduler

view on the picking done to display stock valuation layers

task-1875873