[IMP] l10n_be_hr_payroll_fleet: Allow to set a specific car on payslip
Add a field vehicle_id on the payslip  (pre filled with the company car set on the contract).

The Payroll officer can change it. Use the ATN of this car for the payslip computation.

In Payslip search, add a filter on "Check Car". It filters on payslip where employee is set as driver on the vehicle.

Task: 1925505