[DO NOT MERGE IN STABLE] web: remove the unhandledrejection polyfill
the unhandledrejection polyfill has been added for firefox to show
promises rejected with errors (in odoo, we never reject promises with
error and not cought, so we could assume that if a promise is rejected
with an error, that error must be shown to the user).

However, in the SIP library, there are some promises that are rejected
with an error and handled at a later stage in the process : typically
when requested access to the microphone has been denied. With the polyfill
it shows the error, if we take the polyfill out, the error is correctly
handled by the library.

In the meantime, this polyfill can be overlooked, but do not merge this in master