[IMP] mail: improve author and email from default computation
Purpose of this commit is to correctly compute author_id and email_from
in mail_message as they depends from each other. Previously it was
possible to have desynchronized values. This is improved here by correctly
inheriting from default_get instead of having two default values.

Raise of UserError in default get of email_from is removed as it is not the
role of a default computation to raise an error. Instead it is done manually
at create in mail_message.

Code calling those default methods is adapted in various addons. Generally
it is even better now than before as the raise is not necessary in those
applications, as it is only a default value on a required field.

Related to task 1853147
Linked to PR #32243