[FIX] web: folding extra stages in dropdown
Before this commit, in Project->Task, when you add more stages(so many that it exceeds the width of statusbar)
the layout of statusbar changes and the extra stages are shown on top of older stages because of space issues.

After this commit, there will be fix number of stages allowed (depending upon the device you're using).
The extra stages will be moved to the dropdown of "More".

Task ID:1932190