[IMP] ORM: Add new --migrations-path CLI option
This commit adds a new way to use a migrations scripts folder
whithout needing to symlink it to an hardcoded path.

The folder specified in --migrations-path is then being used by
migration.py to find and execute migrations scripts per module
specified in the -u CLI option.

The folder needs to have the following structure:
- <migrations_path folder>
	- <module1 name>
		- <version1>
			- <script1>
			- <script2>
			- ...
			- <scriptn>
		- <version2>
			- <scripts>
	- <module2 name>
		- <versions>
			- <scripts>
	- ...