[IMP] mail: add user notification pointing to a record
If a message has a res_id or res_model, it will appear in the
corresponding chatter and base it's access rights on this record.

This contrains force the message_notify to set res_id and model to false,
this can be a problem if we want a user notification to appear in the
systray, point to a record, but to be hidden in mail.thread.

This commit adds a new message_type 'user_notification' that should be
managed has a mail_message without res_id and model exept that it will
redirect to a record.