[FIX] hr_holidays: Display approve button when needed
This commit is reverting commit 492764c

Why the original commit?
------------------------
To know if the current user has the rights to approve/delete a leave,
two boolean fields are computed (respectively `can_approve`
and `can_reset`). Associated buttons are shown in the form view according to
the field values.
In the compute methods, the `check_approval_update(state)` method is called.
This method checks various conditions, including `ir.rules` by calling the ORM
method `self.check_access_rule('write')`. If one condition or one `ir.rule`
is not met, an Error is raised.

If the method raises, the user does not have the rights.
The exception is catched and the field `can_approve|reset` is set to False.

However, an unexpected behaviour happens when fields are re-computed
in the context of an onchange. In an onchange, `self.ids` is an empty list.
When checking ir.rules, the `_filter_access_rules()` method is used
to retrieve valid records in self by checking every record against ir.rules.
If some records are not in the valid set, they are considered invalid and an
exception is raised.
But because self.ids is empty, no record is actually checked and therefore
no record is valid.
=> in an onchange it always raises an AccessError and the users is considered
as not having the rights to approve nor refuse a leave as soon as an
onchange has been triggered. Thus associated buttons does not appear.

The original commit fixed the problem by preventing checking the rules
when in an onchange.

Why revert?
-----------
While commit 492764c allowed HR users and Team leaders to approve
after an onchange, it had the unintended effect to
show the "Approve" and "Refuse" buttons to other users.
In summary:

- With commit 492764c: buttons are shown more than necessary
(which is confusing to users who don't actually the rights
to approve)

- Without: buttons are never shown after an onchange (edit and create).
HR users and team leaders have to save the record first before
the buttons appears. This probably an acceptable tradeoff.