[FIX] stock: better error message on UoM write
To avoid compromising stock consistency by careless editing of UoMs,
an error is raised if the operation is risky.
However the check that the operation is risky is too broad, and even includes
the case where an UoM that has just been created cannot be modified.

In that case the workaround is to delete the UoM and create a new one.
The deletion would not be allowed if the UoM is used somewhere,
so that would work only if the operation should have been allowed.

opw 1959679