[FIX] web: update buttons in editable list
In an editable view when editing a row, then apply or remove any groupby or filter the control panel buttons are not updated.

When record is in edit mode in editable listview and applying groupby then control panel buttons 'Save' and 'Discard' buttons are always displaying, this issue occurs because when applying or removing filters contol panel buttons not updated i.e. _updateButtons not called.

With this commit button are updated when _update of list controller is called, _update of list controller is also called from search.

Task-1967278