[IMP] hr_contract: Improve help for contract state
Purpose
=======
Users may not be aware that contract states are
automatically changed by a cron job.
e.g. from Incoming to Running the first day of
a contract.

This might lead to users wanting to set
two contracts in running state for the same employee:
The current contract will expire in two weeks and
a new contract will then begin. But while the HR
is at it, he will set the new contract state to Running
just to be sure not to forget, not knowing it is
automatically managed for him.
Having two contracts in running state might lead to
unexpected behaviours regarding calendars.

Specification
=============
Add a better help message for the state field
explaining that states are automatically managed.

Maybe a hard contraint should be added to avoid
having two running/pending contracts altogether.

While we are at it, add a brief description
for each state.