[IMP] stock: added PDF version in the report of Barcodes
In Inventory, we have an option to print the report of Barcodes in ZPL but we do not have option
to print the same in PDF.

After this commit, there will be a PDF version of the report of Barcodes available just similar
to the report in ZPL.

The two methods added here -> do_print_barcodes_zpl and do_print_barcodes_pdf will be called by
the rpc in methods of stock_barcode module of enterprise so as to provide both the reports
over there.

Also a PDF template is added below the ZPL template.

Task ID: #1970506