[FIX] stock_account: COGS and reconciliation
Commit 98ee15977a3d9d84d made the Stock Interim accounts reconciliable,
and introduced auto-reconciliation of related entries. This has the
advantage to clean-up entries in these accounts. However, since these
accounts are reconciliable, the entries will show up in the manual
reconciliation widget (expected) and the bank reconciliation widget (not
expected).

There are 2 cases where it causes issues:
- after migration from v11, the entries were not reconciliced. Although
  this will be fixed in the near future, it is too late for already
  migrated databases
- using MRP. A list of components i costing Xi might be used to create a
  finished product of cost Y != sum(Xi). Indeed, the routing cost has to
  be taken into account. In this case, the entries are not fully
  reconcilied.

The result is that the bank reconciliation widget ends up with an
growing number of unreconcilied lines, and in the end making it
impossible to use.

In order to work around the issue, we prevent any account used as a
stock interim account to be included in the bank reconciliation widget.
For performance reasons, we directly search the `ir.property` table for
the corresponding accounts instead of filtering AML thanks to
`excluded_ids`.

Closes #32611

opw-1963025
opw-1967462