[FIX] mail: message_unread_counter with channel followers
=== Summary ===

Before this commit, when a channel is followed by another channel
and a new message is posted on this channel, unread notification may
persist indefinitely.

=== Scenario ===

1. Suppose channels:
   - 'original'
   - 'follower'
2. 'follower' is following 'original'
3. New message is posted in 'original'
4. User opens 'original' and sees last message

On page reload, following channel data are fetched:

```
                           original         follower
                        +-------------+-------------------+
 message_unread_counter |     1       |         0         |
 (last) seen_message_id | (last msg)  | (before-last msg) |
                        +-------------+-------------------+
```

`message_unread_counter` is wrong. This is what should be expected:

```
                           original         follower
                        +-------------+-------------------+
 message_unread_counter |     0       |         1         |
 (last) seen_message_id | (last msg)  | (before-last msg) |
                        +-------------+-------------------+
```

Because of this, it tells that 'original' has unread messages.
Marking it as read and reloading page still displays it as unread.
The channel 'follower' is the unread one, but since it is not
displayed as unread, it cannot be marked as read in order to remove
the notification.

As a reminder, seeing a channel consists of setting `seen_message_id`
to the last message of the channel for the current user. This data is
correct, but because unread counter is greater than 0, it tells that
channel is unread.

=== Explanation ===

On page reload, there's RPC on controller `mail/init_messaging`
that fetches data for all channels of current user. If user
happens to be member of both channels (followee and follower),
then it would lead to wrong computation of `message_unread_counter`:

It selects all unread messages from all considered channels and
determines counters of channels based on field `res_id` of the
message, which is the ID of the original channel this message has
been posted. The unread counter is only incremented for the original
channel, but seeing the message on original message would still show
this channel as unread. Indeed, this message is unread on follower
channel, therefore this message's `res_id` also refers to original
channel, thus incrementing the unread counter of original channel.