[REF] point_of_sale: create a PoS picking type for each warehouse
To suggest a default picking type for each newly created PoS Config in a multi-company environment, it is required that all warehouses have a PoS picking type assigned.
As of this commit, there is no longer a default PoS picking type assigned to the first warehouse but a distinct PoS picking type for each warehouse and thus for each PoS Config.

Related to task #1970513