[IMP] web: display Warning and Error Dialog in the website
Some other changes had to be made to display the error/warning modal
in the website.
The dialogs are not loading the template in website so they had to be
extended depending on the type of error.
This has to be done this way because Dialog can load and use xml files
asynchronously using xmlDependancies and templates.

task-1894820