[IMP] website: handle User/Validation Error with beautiful page
All rpc errors/warnings are now displayed using the same modal system as the
backend (crash_manager.js) same goes with User/Validation Errors that would
trigger during a rpc.
For the redirections triggering a User/Validation Error that are currently
displayed as a 500 Error with an ugly white page, the error page will be
similar to 404/403 error but with code 400.
I have also modified the error card on all the Error pages to be loaded on an
open state displaying clearly the reason of the error without clicking anything.

task-1894820