[IMP] mail: to keep the follower preferences dropdown open
Currently, when check/unchecks the follower preferences, a dropdown
closes every time and it is really time-consuming. This is happening
because when click on any follower preferences checkbox, another click
fire on 'tab' from 'setLocalState'.

That's why override 'getLocalState' and 'setLocalState' to keep this
dropdown open until the user clicks out of it.

Currently, qunit is working fine regarding 'dropdown should be opened'
because there is a simple form which is not containing any 'tab'. But
when the form contains any 'tab' then a dropdown is not remaing open.

That's why added 'tab' inside the form to check that dropdown should be
opened in both cases (i.e. with 'tab' also).

task-1960834