[FIX] web: measure/groupby many2one variation in pivot view
In case a many2one is used simultaneously as a
groupby and a measure in the pivot view, the
values returned by a read_group for that
many2one are of the form [id, label] and the
aggregated values (1) are missing in the
read_group results.
As a consequence, a special processing of
read_group results (data points) has to be
done in that situation.
In the comparison mode it was not done
appropriately in case two data points comming
from data and comparison data sets had same
groupby values. For that particular case, the variation of many2one values was not correctly computed. The reason is that the
variation value was not computed using the
right many2one value for the data point in
data set, nor assigned to the right key in
the data point created from the two original
data points.
This is now fixed.